requirements.txt
hf-xet==1.2.0
huggingface==0.0.1
huggingface-hub==0.36.0
matplotlib==3.9.4
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
pillow==11.3.0
scikit-image==0.24.0
scipy==1.13.1
timm==1.0.22
torch==2.8.0
torchmetrics==1.8.2
torchvision==0.23.0
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
scikit-learn==1.6.1
